backend/requirements.txt
fastapi
uvicorn
pillow
cryptography
python-multipart
